implement simple image caching
Only the downloaded image is downloaded and saved locally. Images aren't
resized (yet?) as there probably isn't much to gain by it as Trakt
already resizes into three formats. And QML is most likely capable
enough to efficiently resize the image to the displayed size (due to it
running on QML).